Use `serde(with = ...)`
