+++ src/lib.rs
@@ -10,6 +10,7 @@ pub mod logging;
 pub mod metrics;
 pub mod push;
 pub mod runtime_config;
+pub mod serde;
 pub mod store;
 pub mod test_utils;
 pub mod tokio;
